refactor: moved pages to a separate folder
