+++ package.json
@@ -16,6 +16,7 @@
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.8",
     "react-router-dom": "^6.9.0",
     "sort-by": "^1.2.0"
   },
