+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import "./index.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { MainPage } from "./components/MainPage";
+import { MainPage } from "./pages/MainPage";
 import { Layout } from "./components/Layout";
-import { AboutUs } from "./components/AboutUs";
-import { Page404 } from "./components/Page404";
-import { FormPage } from "./components/FormPage";
+import { AboutUs } from "./pages/AboutUs";
+import { Page404 } from "./pages/Page404";
+import { FormPage } from "./pages/FormPage";
 
 export class App extends React.Component {
   render() {
